requirements.txt
# Core dependencies (existing)
hidapi
mido
python-rtmidi
psutil

# REST API (Phase 3)
fastapi
uvicorn[standard]

# MQTT/Home Assistant (Phase 5)
paho-mqtt
